refactor: move `indexer_sol` and Solana client to a dedicated module
