+++ chain-signatures/node/src/indexer_sol/config.rs
@@ -1,0 +1,24 @@
+use std::fmt;
+
+#[derive(Clone)]
+pub struct SolConfig {
+    /// The solana account secret key used to sign solana respond txn.
+    pub account_sk: String,
+    /// Solana RPC http URL
+    pub rpc_http_url: String,
+    /// Solana RPC websocket URL
+    pub rpc_ws_url: String,
+    /// The program address to watch
+    pub program_address: String,
+}
+
+impl fmt::Debug for SolConfig {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SolConfig")
+            .field("account_sk", &"<hidden>")
+            .field("rpc_http_url", &self.rpc_http_url)
+            .field("rpc_ws_url", &self.rpc_ws_url)
+            .field("program_address", &self.program_address)
+            .finish()
+    }
+}
