+++ chain-signatures/node/src/indexer_sol/mod.rs
@@ -1,11 +1,13 @@
+mod client;
+mod config;
+
 use crate::protocol::Chain;
 use crate::sign_bidirectional::hash_rlp_data;
-use crate::solana_client::{SolanaCatchupBlock, SolanaClient, MAX_CONCURRENT_CHUNK_SIZE};
 use crate::stream::{ChainIndexer, ChainStream};
 use crate::util::ethabi_request_id;
 use crate::util::retry::{retry_async, RetryConfig, RetryError, RetryReason};
-
-pub use crate::solana_client::SolConfig;
+pub use client::{SolanaCatchupBlock, SolanaClient, MAX_CONCURRENT_CHUNK_SIZE};
+pub use config::SolConfig;
 
 use std::collections::{BTreeSet, HashMap, VecDeque};
 use std::pin::Pin;
@@ -55,69 +57,6 @@ const CPI_RESPOND_EVENT_HINTS: &[&str] = &[
     "Program log: Instruction: Respond",
     "Program log: Instruction: RespondBidirectional",
 ];
-
-/// Configures Solana indexer.
-#[derive(Debug, Clone, clap::Parser)]
-#[group(id = "indexer_sol_options")]
-pub struct SolArgs {
-    /// The solana account secret key used to sign solana respond txn.
-    #[arg(long, env("MPC_SOL_ACCOUNT_SK"))]
-    pub sol_account_sk: Option<String>,
-    /// Solana RPC HTTP URL
-    #[clap(long, env("MPC_SOL_RPC_HTTP_URL"), requires = "sol_account_sk")]
-    pub sol_rpc_http_url: Option<String>,
-    /// Solana RPC WS URL
-    #[clap(long, env("MPC_SOL_RPC_WS_URL"), requires = "sol_account_sk")]
-    pub sol_rpc_ws_url: Option<String>,
-    /// The program address to watch
-    #[clap(long, env("MPC_SOL_PROGRAM_ADDRESS"), requires = "sol_account_sk")]
-    pub sol_program_address: Option<String>,
-}
-
-impl SolArgs {
-    pub fn into_str_args(self) -> Vec<String> {
-        let mut args = Vec::with_capacity(6);
-        if let Some(sol_account_sk) = self.sol_account_sk {
-            args.extend(["--sol-account-sk".to_string(), sol_account_sk]);
-        }
-        if let Some(sol_rpc_http_url) = self.sol_rpc_http_url {
-            args.extend(["--sol-rpc-http-url".to_string(), sol_rpc_http_url]);
-        }
-        if let Some(sol_rpc_ws_url) = self.sol_rpc_ws_url {
-            args.extend(["--sol-rpc-ws-url".to_string(), sol_rpc_ws_url]);
-        }
-        if let Some(sol_program_address) = self.sol_program_address {
-            args.extend(["--sol-program-address".to_string(), sol_program_address]);
-        }
-        args
-    }
-
-    pub fn into_config(self) -> Option<SolConfig> {
-        Some(SolConfig {
-            account_sk: self.sol_account_sk?,
-            rpc_http_url: self.sol_rpc_http_url?,
-            rpc_ws_url: self.sol_rpc_ws_url?,
-            program_address: self.sol_program_address?,
-        })
-    }
-
-    pub fn from_config(config: Option<SolConfig>) -> Self {
-        match config {
-            Some(config) => SolArgs {
-                sol_account_sk: Some(config.account_sk),
-                sol_rpc_http_url: Some(config.rpc_http_url),
-                sol_rpc_ws_url: Some(config.rpc_ws_url),
-                sol_program_address: Some(config.program_address),
-            },
-            None => SolArgs {
-                sol_account_sk: None,
-                sol_rpc_http_url: None,
-                sol_rpc_ws_url: None,
-                sol_program_address: None,
-            },
-        }
-    }
-}
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
 pub struct SolSignRequest {
